First pass at breaking out controllers
